refactor: update `generatorArgs` method in project generator
Changed the generatorArgs method in QwikCliProjectGenerator to improve readability and simplify the arguments passed. Instead of asynchronously retrieving the project name, it now receives it directly from the project parameter. In addition, the temporary directory creation redundancy was removed, and a fixed string "basic" is now used to keep the directory structure more straightforward.